# webapp/frontend/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ConfigDetail": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parse_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parse Error"
          },
          "parsed": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsed"
          },
          "raw": {
            "title": "Raw",
            "type": "string"
          }
        },
        "required": [
          "name",
          "raw"
        ],
        "title": "ConfigDetail",
        "type": "object"
      },
      "ConfigEntry": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ConfigEntry",
        "type": "object"
      },
      "ConfigKind": {
        "enum": [
          "experiment",
          "universe",
          "study",
          "strategy",
          "hpo"
        ],
        "title": "ConfigKind",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StudySpecUploadCreate": {
        "description": "POST /configs/study/uploads body.",
        "properties": {
          "slug": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9_\\-]*$",
            "title": "Slug",
            "type": "string"
          },
          "yaml": {
            "maxLength": 131072,
            "minLength": 1,
            "title": "Yaml",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "yaml"
        ],
        "title": "StudySpecUploadCreate",
        "type": "object"
      },
      "StudySpecUploadDetail": {
        "description": "Single-upload read shape with the full YAML body.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "owner_user_id": {
            "title": "Owner User Id",
            "type": "integer"
          },
          "owner_username": {
            "title": "Owner Username",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "yaml": {
            "title": "Yaml",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "yaml",
          "created_at",
          "updated_at",
          "owner_user_id",
          "owner_username"
        ],
        "title": "StudySpecUploadDetail",
        "type": "object"
      },
      "StudySpecUploadSummary": {
        "description": "Listing entry - no YAML body, no validation errors.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "owner_user_id": {
            "title": "Owner User Id",
            "type": "integer"
          },
          "owner_username": {
            "title": "Owner Username",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "created_at",
          "updated_at",
          "owner_user_id",
          "owner_username"
        ],
        "title": "StudySpecUploadSummary",
        "type": "object"
      },
      "StudySpecValidateRequest": {
        "description": "``POST /configs/study_spec/validate`` body - raw YAML text.\n\nA dedicated wire type keeps the existing ValidateRequest (parsed payload)\nuncoupled from the YAML-text path, which has to handle parse errors before\nPydantic ever sees the body.",
        "properties": {
          "yaml": {
            "title": "Yaml",
            "type": "string"
          }
        },
        "required": [
          "yaml"
        ],
        "title": "StudySpecValidateRequest",
        "type": "object"
      },
...
      "UniverseSpecUploadCreate": {
        "description": "POST /configs/universe/uploads body.",
        "properties": {
          "slug": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9_\\-]*$",
            "title": "Slug",
            "type": "string"
          },
          "yaml": {
            "maxLength": 131072,
            "minLength": 1,
            "title": "Yaml",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "yaml"
        ],
        "title": "UniverseSpecUploadCreate",
        "type": "object"
      },
      "UniverseSpecUploadDetail": {
        "description": "Single-upload read shape with the full YAML body.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "owner_user_id": {
            "title": "Owner User Id",
            "type": "integer"
          },
          "owner_username": {
            "title": "Owner Username",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "yaml": {
            "title": "Yaml",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "yaml",
          "created_at",
          "updated_at",
          "owner_user_id",
          "owner_username"
        ],
        "title": "UniverseSpecUploadDetail",
        "type": "object"
      },
      "UniverseSpecUploadSummary": {
        "description": "Listing entry - no YAML body.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "owner_user_id": {
            "title": "Owner User Id",
            "type": "integer"
          },
          "owner_username": {
            "title": "Owner Username",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "created_at",
          "updated_at",
          "owner_user_id",
          "owner_username"
        ],
        "title": "UniverseSpecUploadSummary",
        "type": "object"
      },
      "UniverseSpecValidateRequest": {
        "description": "``POST /configs/universe_spec/validate`` body - raw YAML text.",
        "properties": {
          "yaml": {
            "title": "Yaml",
            "type": "string"
          }
        },
        "required": [
          "yaml"
        ],
        "title": "UniverseSpecValidateRequest",
        "type": "object"
      },
...
      "ValidateRequest": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ConfigKind"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "kind",
          "payload"
        ],
        "title": "ValidateRequest",
        "type": "object"
      },
      "ValidateResponse": {
        "properties": {
          "errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            },
            "title": "Errors",
            "type": "array"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid"
        ],
        "title": "ValidateResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationErrorItem": {
        "properties": {
          "loc": {
            "items": {
              "type": "string"
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationErrorItem",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Quant Webapp",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/configs/study/uploads": {
      "get": {
        "operationId": "get_study_uploads_api_configs_study_uploads_get",
        "parameters": [
          {
            "in": "query",
            "name": "all",
            "required": false,
            "schema": {
              "default": false,
              "title": "All",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StudySpecUploadSummary"
                  },
                  "title": "Response Get Study Uploads Api Configs Study Uploads Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Study Uploads",
        "tags": [
          "configs"
        ]
      },
      "post": {
        "operationId": "post_study_upload_api_configs_study_uploads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudySpecUploadCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudySpecUploadDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Study Upload",
        "tags": [
          "configs"
        ]
      }
    },
    "/api/configs/study/uploads/{slug}": {
      "delete": {
        "operationId": "delete_study_upload_api_configs_study_uploads__slug__delete",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Study Upload",
        "tags": [
          "configs"
        ]
      },
      "get": {
        "operationId": "get_study_upload_api_configs_study_uploads__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudySpecUploadDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Study Upload",
        "tags": [
          "configs"
        ]
      }
    },
...
    "/api/configs/study_spec/validate": {
      "post": {
        "operationId": "post_validate_study_spec_api_configs_study_spec_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudySpecValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Validate Study Spec",
        "tags": [
          "configs"
        ]
      }
    },
    "/api/configs/universe/uploads": {
      "get": {
        "operationId": "get_universe_uploads_api_configs_universe_uploads_get",
        "parameters": [
          {
            "in": "query",
            "name": "all",
            "required": false,
            "schema": {
              "default": false,
              "title": "All",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UniverseSpecUploadSummary"
                  },
                  "title": "Response Get Universe Uploads Api Configs Universe Uploads Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Universe Uploads",
        "tags": [
          "configs"
        ]
      },
      "post": {
        "operationId": "post_universe_upload_api_configs_universe_uploads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UniverseSpecUploadCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniverseSpecUploadDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Universe Upload",
        "tags": [
          "configs"
        ]
      }
    },
    "/api/configs/universe/uploads/{slug}": {
      "delete": {
        "operationId": "delete_universe_upload_api_configs_universe_uploads__slug__delete",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Universe Upload",
        "tags": [
          "configs"
        ]
      },
      "get": {
        "operationId": "get_universe_upload_detail_api_configs_universe_uploads__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniverseSpecUploadDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Universe Upload Detail",
        "tags": [
          "configs"
        ]
      }
    },
...
    "/api/configs/universe_spec/validate": {
      "post": {
        "operationId": "post_validate_universe_spec_api_configs_universe_spec_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UniverseSpecValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Validate Universe Spec",
        "tags": [
          "configs"
        ]
      }
    },
    "/api/configs/validate": {
      "post": {
        "operationId": "post_validate_api_configs_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Validate",
        "tags": [
          "configs"
        ]
      }
    },
    "/api/configs/{kind}": {
      "get": {
        "operationId": "get_configs_api_configs__kind__get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ConfigKind"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConfigEntry"
                  },
                  "title": "Response Get Configs Api Configs  Kind  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Configs",
        "tags": [
          "configs"
        ]
      }
    },
    "/api/configs/{kind}/{name}": {
      "get": {
        "operationId": "get_config_detail_api_configs__kind___name__get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ConfigKind"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Config Detail",
        "tags": [
          "configs"
        ]
      }
    },
...
  }
}